Split the code into blocks to improve readability.
