+++ rl_algorithms/utils/replaybuffer.py
@@ -1,0 +1,20 @@
+import random
+from collections import deque
+
+
+class ReplayBuffer:
+    """
+    ReplayBuffer，用于存储经验
+    """
+
+    def __init__(self, max_size):
+        self.buffer = deque(maxlen=max_size)
+
+    def add(self, state, action, reward, next_state, done):
+        self.buffer.append((state, action, reward, next_state, done))
+
+    def sample(self, batch_size):
+        return random.sample(self.buffer, batch_size)
+
+    def __len__(self):
+        return len(self.buffer)
